Patch: dynamic-parserdata Type: bugfix
In the original SMW, ParserData objects for function hooks are initialised one time
at the moment of setting the hook. This is incorrect because ParserData contains
a reference to Title and ParserOutput and they are changed during the Parser's
life when clearState() is called, basically on every subsequent parse() call
with $clearState == true.

This, for example, prevents to retrieve SemanticData from ParserOutput during
prepareContentForEdit() before saving a revision.