Patch: synchronous-cache-purge-jobs Type: bugfix
Process SMW's ParserCachePurgeJobs synchronously when $smwgEnabledHttpDeferredJobRequest is off.
This should fix "non-immediate" flushing of dependent queries on page updates.

By the way, DeferredRequestDispatchManager is a bad idea in whole, because with PHP's session
locking it leads to deadlocks in apache/php-fpm:
- initial request locks sessions and posts a 'loopback' request to run jobs
- loopback request also tries to lock the same session and hangs
- initial request handler also hangs because it waits for loopback to complete
- other requests from the same user also hang because they also want the same session
- bang, mediawiki is out.

To mitigate this, $smwgEnabledHttpDeferredJobRequest must be turned off, or 'no-session-lock'
patch from Mediawiki4Intranet should be used.